[CHORE#1606] 파일들 디렉토리 경로 변경
